// src/interfaces/form.ts
export interface MyFormData {
  email?: string;
  password?: string;
  confirm?: string;
  policy?: boolean;
  firstName?: string;
  lastName?: string;
  phone?: string;
  altPhone?: string;
  cpf?: string;
  petType?: string;
  petName?: string;
  petBreed?: string;
  petGender?: string;
  petBirthDay?: string;
  petSpayedOrNeutered?: string;
  petWeight?: string;
  petPhoto?: string;

  kisses?: boolean;
  walk?: boolean;
  barking?: boolean;
  snuggling?: boolean;
  treats?: boolean;
  playingFetch?: boolean;
  naps?: boolean;
  toys?: boolean;

  petDetail?: string;
}